server/worker: add webhook queue to observability metrics
